refactor(db): extract db-builtin-skills and db-migrations
Split lib/db.ts into focused modules:
- lib/db-builtin-skills.ts: BUILTIN_AGENT_BROWSER_SKILL constant,
  deriveSkillDescription helper
- lib/db-migrations.ts: migrate function, backfillVisionMcpServers
- lib/db.ts retains getDb, resetDbForTests, getDatabasePath with
  re-exports for backward compatibility
